requirements.txt
flask
supabase
python-dotenv
bcrypt
gunicorn